requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
click==8.2.2
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflect==7.5.0
iniconfig==2.1.0
isort==6.0.1
more-itertools==10.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymssql==2.3.7
pytest==8.4.1
python-dotenv==1.1.1
python-jose==3.5.0
requests
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.42
starlette==0.47.2
typeguard==4.4.4
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
